Fix Redcarpet engine options
The engine uses the "old" option names, created for compatibility with redcloth.
This change removes the `gh_blockcode` option, which has no effect,
and changes `fenced_code` to `fenced_code_blocks`, which is the "new" way
of specifying the same option.

Here is some sample markdown that renders correctly with this change:

	Here is a fenced code block:

	~~~
	code in a fenced block
	~~~

	And another fenced code block, which shouldn't be red:

	```
	another code block with fences
	```

	The word `ruby` should not be visible in this code block:

	```ruby
	fence ALL the code blocks!
	```